Update to fix negative diffusivity issue
system eigen must be 'n' i.e. the git-shipped version of eigen must be used. This must be used in conjunction with Eigen3.2.9 for CharlesX in order for finite rate chemistry to work. IF you do not do this, the code may compile, may run, and may even give results but there will be GARBAGE in the results due to miscalculation of the diffusive quantities by Cantera.